remove mocks from vr-object tests, don't use globals for helpers
